First merge with branch V1.2BETA

+++ src/ec/nodes/EcAddOperator.java
@@ -8,7 +8,7 @@ package ec.nodes;
 public class EcAddOperator extends EcOperator {
 
 	@Override
-	public double calculateOutput(int input) {
+	public double calculateOutput(double input) {
 		double output = 0;
 		output += (leftChild.calculateOutput(input) + rightChild.calculateOutput(input));
 		return output;
